formatting code

+++ src/main/java/ru/geekbrains/videoservice/services/AmazonService.java
@@ -9,13 +9,10 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
-import io.github.techgnious.IVCompressor;
-import io.github.techgnious.dto.ResizeResolution;
-import io.github.techgnious.dto.VideoFormats;
-import io.github.techgnious.exception.VideoException;
+import org.apache.commons.io.FilenameUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import ru.geekbrains.videoservice.Const.AmazonConst;
+import ru.geekbrains.videoservice.constants.AmazonConst;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -30,23 +27,11 @@ import java.util.List;
 @Service
 public class AmazonService {
 
-//    @Value("${amazon.access-key}")
-//    private String accessKey;
-//    @Value("${amazon.secret-key}")
-//    private String secretKey;
-//    @Value("${amazon.server-endpoint}")
-//    private String serverEndpoint;
-//    @Value("${amazon.region}")
-//    private String region;
-//    @Value("${amazon.bucket}")
-//    private String bucket;
-//    @Value("${amazon.storage-url")
-//    private String storageUrl;
-
     AWSCredentials credentials = new BasicAWSCredentials(
             AmazonConst.ACCESS_KEY,
             AmazonConst.SECRET_KEY
     );
+
 
     AmazonS3 s3 = AmazonS3ClientBuilder.standard()
             .withCredentials(new AWSStaticCredentialsProvider(credentials))
@@ -55,8 +40,9 @@ public class AmazonService {
             )
             .build();
 
+
     public List<String> getAllFiles() {
-        List<String> filesNames = new ArrayList<>();
+        List<String> filesNames = new ArrayList();
         List<S3ObjectSummary> files = s3.listObjects(AmazonConst.BUCKET).getObjectSummaries();
         for (S3ObjectSummary list : files) {
             filesNames.add(AmazonConst.STORAGE_URL + list.getKey());
@@ -64,22 +50,26 @@ public class AmazonService {
         return filesNames;
     }
 
-    public void uploadFile(MultipartFile file) throws IOException, VideoException {
+
+    public void uploadFile(MultipartFile file) {
+        String extension = FilenameUtils.getExtension(file.getOriginalFilename());
+        if (!extension.equals("mp4")) {
+            throw new IllegalArgumentException("File format should be mp4");
+        }
         File fileObj = convertMultiPartFileToFile(file);
-        s3.putObject(new PutObjectRequest(AmazonConst.BUCKET, fileObj.getName(), fileObj));
+        String fileName = file.getOriginalFilename();
+        s3.putObject(new PutObjectRequest(AmazonConst.BUCKET, fileName, fileObj));
         fileObj.delete();
     }
 
-    private File convertMultiPartFileToFile(MultipartFile file) throws IOException, VideoException {
-        File compressedFile = new File(String.valueOf(System.currentTimeMillis()) + ".mp4");
-        IVCompressor compressor = new IVCompressor();
-        byte[] compressed = compressor.reduceVideoSize(file.getBytes(), VideoFormats.MP4, ResizeResolution.R720P);
-        try (FileOutputStream fos = new FileOutputStream(compressedFile)) {
-            fos.write(compressed);
+    private File convertMultiPartFileToFile(MultipartFile file) {
+        File convertedFile = new File(file.getOriginalFilename());
+        try (FileOutputStream fos = new FileOutputStream(convertedFile)) {
+            fos.write(file.getBytes());
         } catch (IOException e) {
             e.printStackTrace();
         }
-        return compressedFile;
+        return convertedFile;
     }
 
 }
